Rename files to use UpperCamelCaseCluster/Logic for time-format-localization-server
